# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Codalla AI Coding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflow presets
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}:
    put:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow preset
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow preset
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    WorkflowStep:
      type: object
      required: [title, prompt]
      properties:
        title:
          type: string
        prompt:
          type: string
    Workflow:
      type: object
      required: [id, name, steps, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        isShared:
          type: boolean
        isOwner:
          type: boolean
        description:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkflowInput:
      type: object
      required: [name, steps]
      properties:
        name:
          type: string
        isShared:
          type: boolean
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
...