# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Codalla AI Coding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowStep } from './workflowStep';

export interface Workflow {
  id: string;
  name: string;
  isShared?: boolean;
  isOwner?: boolean;
  description?: string | null;
  steps: WorkflowStep[];
  createdAt: string;
  updatedAt: string;
}