# lib/api-zod/src/generated/types/workflowInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Codalla AI Coding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowStep } from './workflowStep';

export interface WorkflowInput {
  name: string;
  isShared?: boolean;
  description?: string;
  steps: WorkflowStep[];
}